pyproject.toml
[project]
name = "elspeth"
version = "0.1.0"
description = "Auditable Sense/Decide/Act pipelines for high-reliability systems"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "John" }]
keywords = ["elspeth", "audit", "pipeline", "sda", "traceability", "data-processing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # CLI and TUI
    "typer>=0.12",
    "textual>=0.52",

    # Configuration
    "dynaconf>=3.2",
    "pydantic>=2.6",

    # Plugin system
    "pluggy>=1.4",

    # Data handling
    "pandas>=2.2",

    # HTTP client
    "httpx>=0.27",

    # Database
    "sqlalchemy>=2.0",
    "alembic>=1.13",

    # Retries
    "tenacity>=8.2",

    # === Acceleration Stack ===
    # Canonical JSON (RFC 8785/JCS)
    "rfc8785>=0.1",

    # DAG validation
    "networkx>=3.2",

    # Observability
    "opentelemetry-api>=1.23",
    "opentelemetry-sdk>=1.23",
    "opentelemetry-exporter-otlp>=1.23",

    # Structured logging
    "structlog>=24.1",

    # Rate limiting
    "pyrate-limiter>=3.1",

    # Diffing (for verify mode)
    "deepdiff>=7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
    "hypothesis>=6.98",
    "mypy>=1.8",
    "ruff>=0.3",
    "pre-commit>=3.6",
    # Type stubs
    "pandas-stubs>=2.2",
    "types-networkx>=3.2",
    # Mutation testing (Phase 3)
    "mutmut>=2.4",
]

llm = [
    # Phase 6: LLM plugin pack
    "jinja2>=3.1",
    "httpx>=0.27",
    "openai>=1.0",
    "litellm>=1.30",
]

azure = [
    # Phase 7: Azure plugin pack
    "azure-storage-blob>=12.19",
    "azure-identity>=1.15",
    "jinja2>=3.1",  # Used by AzureBlobSink for path templating
]

all = [
    "elspeth[dev,llm,azure]",
]

[project.scripts]
elspeth = "elspeth.cli:app"
check-contracts = "scripts.check_contracts:main"

[project.urls]
Homepage = "https://github.com/your-org/elspeth"
Documentation = "https://github.com/your-org/elspeth#readme"
Repository = "https://github.com/your-org/elspeth"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/elspeth"]

# === Ruff Configuration ===
[tool.ruff]
target-version = "py311"
line-length = 140
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "W",      # pycodestyle warnings
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez (important for audit timestamps!)
    "T20",    # flake8-print (no stray prints in production)
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "SIM108", # ternary operator - prefer explicit if/else for readability with comments
]

[tool.ruff.lint.isort]
known-first-party = ["elspeth"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["T20"]  # Allow print in tests
"scripts/**/*.py" = ["T20"]  # Allow print in CLI scripts
"alembic/**/*.py" = ["E402"]  # Alembic env.py has imports after config setup
"vulture_whitelist.py" = ["B018", "F821"]  # Vulture whitelist uses bare names
# Package __init__ files use categorical __all__ grouping for readability
"src/elspeth/contracts/__init__.py" = ["RUF022"]
"src/elspeth/plugins/__init__.py" = ["RUF022"]

# === Mypy Configuration ===
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_unreachable = true
show_error_codes = true
plugins = ["pydantic.mypy"]
# Allow unused ignores - needed for pre-commit which runs in isolated env
warn_unused_ignores = false

[[tool.mypy.overrides]]
module = [
    "dynaconf.*",
    "pluggy.*",
    "rfc8785.*",
    "pyrate_limiter.*",
    "deepdiff.*",
    "textual.*",
    "opentelemetry.*",
    "hypothesis.*",
    "azure.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
# Test files use hypothesis decorators that don't have type stubs
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = "elspeth.tui.*"
# TUI uses textual which has incomplete type stubs
# App[None] inheritance triggers "Class cannot subclass value of type Any"
disallow_subclassing_any = false

# === Pytest Configuration ===
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external services",
    "asyncio: marks tests as async (handled by pytest-asyncio)",
]

# === Coverage Configuration ===
[tool.coverage.run]
source = ["src/elspeth"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]

# === Mutation Testing Configuration (mutmut 2.x) ===
[tool.mutmut]
# Paths to mutate (start with critical audit paths)
paths_to_mutate = "src/elspeth/core/"
# Test runner command - run property tests and core tests with fast-fail
runner = "python -m pytest tests/property/ tests/core/ -x --tb=no -q"
# Tests directory
tests_dir = "tests/"
# Disable backup files - we use git
backup = false
